"Ignore" source directories in cobertura reports
Cobertura doesn't keep the full path to analyzed files in its XML
reports, only a list of source directories which were used to find those
files. Before this change, the plugin tried to reconstruct the absolute
path from those informations and tries to "ask" SonarQube about each of
them. Now, we just try to find any input file matching the relative path
of the current file. This makes the logic much simpler and gives us a
slight chance to detect multiple files matching the same relative path.

As a side-effect, we don't need to mock the whole file system in the
test cases.